# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend Developer Node.js Study Lab API
  version: 1.1.0
  description: >-
    Complete OpenAPI reference for the backend study cases.
    Includes request body schemas, request examples, and response examples
    for both success and common error outcomes.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /cases/17-event-sourcing-cqrs/accounts:
    post:
      tags: [Reliability and Domain]
      summary: Open account and append first event
      operationId: openCase17Account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAccountRequest"
            examples:
              openAccount:
                value:
                  accountId: acct-test-1
                  initialBalance: 250
                  commandId: cmd-open-acct-test-1
      responses:
        "201":
          description: Account opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                created:
                  value:
                    case: 17-event-sourcing-cqrs
                    status: accepted
                    event:
                      sequence: 1
                      accountId: acct-test-1
                      commandId: cmd-open-acct-test-1
                      type: account-opened
                      amount: 250
                      occurredAt: 2026-04-11T09:45:00.000Z
                    projection:
                      accountId: acct-test-1
                      balance: 250
                      version: 1
                      updatedAt: 2026-04-11T09:45:00.000Z
        "200":
          description: Duplicate command accepted idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                duplicate:
                  value:
                    case: 17-event-sourcing-cqrs
                    status: duplicate_command
                    commandId: cmd-open-acct-test-1
                    eventSequence: 1
                    projection:
                      accountId: acct-test-1
                      balance: 250
                      version: 1
                      updatedAt: 2026-04-11T09:45:00.000Z
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                validationError:
                  value:
                    case: 17-event-sourcing-cqrs
                    error:
                      fieldErrors:
                        initialBalance:
                          - Number must be greater than or equal to 0
                      formErrors: []
        "409":
          description: Account already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                alreadyExists:
                  value:
                    case: 17-event-sourcing-cqrs
                    error: "account already exists: acct-test-1"
  /cases/17-event-sourcing-cqrs/accounts/{accountId}/deposit:
    post:
      tags: [Reliability and Domain]
      summary: Deposit funds into account
      operationId: depositCase17Funds
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountCommandRequest"
            examples:
              deposit:
                value:
                  amount: 40
                  commandId: cmd-deposit-1
      responses:
        "201":
          description: Deposit event accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                created:
                  value:
                    case: 17-event-sourcing-cqrs
                    status: accepted
                    event:
                      sequence: 2
                      accountId: acct-test-1
                      commandId: cmd-deposit-1
                      type: money-deposited
                      amount: 40
                      occurredAt: 2026-04-11T09:46:00.000Z
                    projection:
                      accountId: acct-test-1
                      balance: 290
                      version: 2
                      updatedAt: 2026-04-11T09:46:00.000Z
        "200":
          description: Duplicate command accepted idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                duplicate:
                  value:
                    case: 17-event-sourcing-cqrs
                    status: duplicate_command
                    commandId: cmd-deposit-1
                    eventSequence: 2
                    projection:
                      accountId: acct-test-1
                      balance: 290
                      version: 2
                      updatedAt: 2026-04-11T09:46:00.000Z
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                invalidAmount:
                  value:
                    case: 17-event-sourcing-cqrs
                    error:
                      fieldErrors:
                        amount:
                          - Number must be greater than 0
                      formErrors: []
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    case: 17-event-sourcing-cqrs
                    error: "account not found: acct-missing"
  /cases/17-event-sourcing-cqrs/accounts/{accountId}/withdraw:
    post:
      tags: [Reliability and Domain]
      summary: Withdraw funds from account
      operationId: withdrawCase17Funds
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountCommandRequest"
            examples:
              withdraw:
                value:
                  amount: 35
                  commandId: cmd-withdraw-1
      responses:
        "201":
          description: Withdrawal event accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                created:
                  value:
                    case: 17-event-sourcing-cqrs
                    status: accepted
                    event:
                      sequence: 3
                      accountId: acct-test-1
                      commandId: cmd-withdraw-1
                      type: money-withdrawn
                      amount: 35
                      occurredAt: 2026-04-11T09:47:00.000Z
                    projection:
                      accountId: acct-test-1
                      balance: 255
                      version: 3
                      updatedAt: 2026-04-11T09:47:00.000Z
        "200":
          description: Duplicate command accepted idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                duplicate:
                  value:
                    case: 17-event-sourcing-cqrs
                    status: duplicate_command
                    commandId: cmd-withdraw-1
                    eventSequence: 3
                    projection:
                      accountId: acct-test-1
                      balance: 255
                      version: 3
                      updatedAt: 2026-04-11T09:47:00.000Z
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                invalidAmount:
                  value:
                    case: 17-event-sourcing-cqrs
                    error:
                      fieldErrors:
                        amount:
                          - Number must be greater than 0
                      formErrors: []
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    case: 17-event-sourcing-cqrs
                    error: "account not found: acct-missing"
        "409":
          description: Insufficient account balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                insufficient:
                  value:
                    case: 17-event-sourcing-cqrs
                    error: insufficient balance
                    projection:
                      accountId: acct-test-1
                      balance: 10
                      version: 6
                      updatedAt: 2026-04-11T09:48:00.000Z
  /cases/17-event-sourcing-cqrs/accounts/{accountId}:
    get:
      tags: [Reliability and Domain]
      summary: Query account projection
      operationId: getCase17Account
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Account projection and event count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 17-event-sourcing-cqrs
                    projection:
                      accountId: acct-test-1
                      balance: 255
                      version: 3
                      updatedAt: 2026-04-11T09:47:00.000Z
                    totalEvents: 3
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    case: 17-event-sourcing-cqrs
                    error: "account not found: acct-missing"
  /cases/17-event-sourcing-cqrs/events:
    get:
      tags: [Reliability and Domain]
      summary: Query all events or filter by accountId
      operationId: listCase17Events
      parameters:
        - in: query
          name: accountId
          required: false
          schema:
            type: string
          description: Optional projection filter
          example: acct-test-1
      responses:
        "200":
          description: Event stream snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
              examples:
                success:
                  value:
                    case: 17-event-sourcing-cqrs
                    count: 3
                    data:
                      - sequence: 1
                        accountId: acct-test-1
                        commandId: cmd-open-acct-test-1
                        type: account-opened
                        amount: 250
                        occurredAt: 2026-04-11T09:45:00.000Z
                      - sequence: 2
                        accountId: acct-test-1
                        commandId: cmd-deposit-1
                        type: money-deposited
                        amount: 40
                        occurredAt: 2026-04-11T09:46:00.000Z
                      - sequence: 3
                        accountId: acct-test-1
                        commandId: cmd-withdraw-1
                        type: money-withdrawn
                        amount: 35
                        occurredAt: 2026-04-11T09:47:00.000Z
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AccountId:
      in: path
      name: accountId
      required: true
      schema:
        type: string
      description: Account identifier
      example: acct-test-1
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
      required: [error]
...
    OpenAccountRequest:
      type: object
      properties:
        accountId:
          type: string
        initialBalance:
          type: number
          minimum: 0
          maximum: 1000000
        commandId:
          type: string
      required: [initialBalance]
    AmountCommandRequest:
      type: object
      properties:
        amount:
          type: number
          exclusiveMinimum: 0
          maximum: 1000000
        commandId:
          type: string
      required: [amount]